requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
streamlit==1.41.1
langchain==0.3.16
langchain-community==0.3.16
langchain-text-splitters==0.3.4
faiss-cpu==1.13.2
sentence-transformers==3.3.1
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.1.0
pdfplumber==0.11.4
python-docx==1.1.2
docx2txt==0.8
rank-bm25==0.2.2
torch==2.5.1+cu124
transformers==4.47.1
numpy==2.3.4
pandas==2.3.3